+++ pkg/filter/stream/common/rengine.go
@@ -16,15 +16,17 @@
 package common
 
 import (
+	"strike/pkg/filter/stream/common/limit"
 	"strike/pkg/filter/stream/common/model"
+	"strike/pkg/filter/stream/common/resource"
 	"strike/pkg/protocol"
 )
 
 // RuleEngine as
 type RuleEngine struct {
 	ruleConfig    *model.RuleConfig
-	//matcherEngine resource.MatcherEngine
-	//limitEngine   *limit.LimitEngine
+	matcherEngine resource.MatcherEngine
+	limitEngine   *limit.LimitEngine
 	//stat          *metrix.Stat
 }
 
@@ -34,35 +36,34 @@ func NewRuleEngine(config *model.RuleConfig) *RuleEngine {
 		ruleConfig: config,
 	}
 
-	//limitEngine, err := limit.NewLimitEngine(config)
-	//if err != nil {
-	//	return nil
-	//}
-	//ruleEngine.limitEngine = limitEngine
-	//ruleEngine.matcherEngine = resource.NewMatcherEnine(nil)
+	limitEngine, err := limit.NewLimitEngine(config)
+	if err != nil {
+		return nil
+	}
+	ruleEngine.limitEngine = limitEngine
+	ruleEngine.matcherEngine = resource.NewMatcherEnine(nil)
 	//ruleEngine.stat = metrix.NewStat(config)
 	return ruleEngine
 }
 
 func (e *RuleEngine) invoke(headers protocol.HeaderMap) bool {
-	//if e.match(headers) {
-	//	e.stat.Counter(metrix.INVOKE).Inc(1)
-	//	if e.limitEngine.OverLimit() {
-	//		e.stat.Counter(metrix.BLOCK).Inc(1)
-	//		if e.ruleConfig.RunMode == model.RunModeControl {
-	//			return false
-	//		}
-	//	}
-	//}
+	if e.match(headers) {
+		//e.stat.Counter(metrix.INVOKE).Inc(1)
+		if e.limitEngine.OverLimit() {
+			//e.stat.Counter(metrix.BLOCK).Inc(1)
+			if e.ruleConfig.RunMode == model.RunModeControl {
+				return false
+			}
+		}
+	}
 	return true
 }
 
-func (e RuleEngine) match(headers protocol.HeaderMap) bool {
-	//return e.matcherEngine.Match(headers, e.ruleConfig)
-	return true
+func (e *RuleEngine) match(headers protocol.HeaderMap) bool {
+	return e.matcherEngine.Match(headers, e.ruleConfig)
 }
 
 //stop timer
-func (e RuleEngine) stop() {
+func (e *RuleEngine) stop() {
 	//e.stat.Stop()
 }